% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_project.R
\name{analyse_project}
\alias{analyse_project}
\title{analyse_project}
\usage{
analyse_project(path, rules = default_rules())
}
\arguments{
\item{path}{}

\item{rules}{}
}
\value{
smells - a object of the detected code smells
}
\description{
analyse_project
}
\examples{
  analyse_project(path)
}
\author{
Mark London
}
