% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_context.R
\name{build_context}
\alias{build_context}
\title{build_context}
\usage{
build_context(parsed)
}
\arguments{
\item{parsed}{The output of parse()}
}
\value{
A list containing a 'calls' tibble.
}
\description{
Extracts all function calls and symbols from parsed R code to build a searchable context.
}
\examples{
  build_context(parsed)
}
\author{
Mark London
}
