% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_audits.R
\name{compare_audits}
\alias{compare_audits}
\title{Compare Two Audit Snapshots}
\usage{
compare_audits(old_path, new_path)
}
\arguments{
\item{old_path}{Path to the older .rds snapshot file.}

\item{new_path}{Path to the newer .rds snapshot file.}
}
\value{
NULL (Prints report to console).
}
\description{
Loads two previously saved audit snapshots and calculates the difference
in key metrics (Misplaced functions, Ghost dependencies, etc.).
}
\author{
Mark London
}
