% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_rules.R
\name{default_rules}
\alias{default_rules}
\title{default_rules}
\usage{
default_rules()
}
\value{
Invisibly returns NULL
}
\description{
default_rules
}
\examples{
  default_rules()
}
\author{
Mark London
}
