% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_smell.R
\name{new_smell}
\alias{new_smell}
\title{new_smell}
\usage{
new_smell(file, line, id, message, severity, category, confidence = "MEDIUM")
}
\arguments{
\item{file}{}

\item{line}{}

\item{id}{}

\item{message}{}

\item{severity}{}

\item{category}{}

\item{confidence}{}
}
\value{
out a class of smell
}
\description{
new_smell
}
\author{
Mark London
}
