% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_dynamic_execution.R
\name{rule_dynamic_execution}
\alias{rule_dynamic_execution}
\title{rule_dynamic_execution}
\usage{
rule_dynamic_execution(ctx)
}
\arguments{
\item{ctx}{}
}
\value{
Invisibly returns NULL
}
\description{
rule_dynamic_execution
}
\examples{
  rule_dynamic_execution(ctx)
}
\author{
Mark London
}
