% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rules.R
\name{run_rules}
\alias{run_rules}
\title{run_rules}
\usage{
run_rules(ctx, rules)
}
\arguments{
\item{rule}{}
}
\value{
dataframe of rules
}
\description{
run_rules
}
\examples{
  run_rules(ctx, rules)
}
\author{
Mark London
}
