% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_context.R
\name{build_context}
\alias{build_context}
\title{build_context}
\usage{
build_context(parsed)
}
\arguments{
\item{parsed}{}
}
\value{
list
}
\description{
build_context
}
\examples{
  build_context(parsed)
}
\author{
Mark London
}
